finish npm task
