--- js-npm-task/index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Document</title>
-  <link rel="stylesheet" href="style.css">
-</head>
-<body>
-  <div class="switch-lang">
-    <label id="switch" class="switch">
-      <input type="checkbox" id="slider">
-      <span class="slider round"></span>
-    </label>
-  </div>
-  <div class="time">
-
-  </div>
-  <script src="index.js"></script>
-</body>
-</html>
