--- js-npm-task/index.js
@@ -1,121 +1,0 @@
-const i18n = require('i18next');
-
-const MONTH_NAME = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December' ];/* ['Январь', 'Февраль', 'Март', 'Апрель', 'Май', 'Июнь', 'Июль', 'Август', 'Сентябрь', 'Октябрь', 'Ноябрь', 'Декабрь'] */
-const WEEK_DAY_NAME = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'];
-const time = document.querySelector('.time');
-const sliderLang = document.getElementById('slider');
-const addZero = (n) => (parseInt(n, 10) < 10 ? '0' : '') + n;
-
-const showTime = () => {
-  const today = new Date();
-  const hour = today.getHours();
-  const min = today.getMinutes();
-  const sec = today.getSeconds();
-  const month = i18next.t(MONTH_NAME[today.getMonth()]);
-  const day = today.getDate();
-  const weekDay = i18next.t(WEEK_DAY_NAME[today.getDay()]);
-  time.innerHTML = `${addZero(hour)}<span>:</span>${addZero(min)}<span>:</span>${addZero(sec)}`;
-  if (i18next.language === 'en') {
-    week.innerHTML = `${weekDay}, ${month} ${day}`;
-  } else {
-    week.innerHTML = ` ${weekDay}, ${day} ${month}`;
-  }
-  setTimeout(showTime, 1000);
-};
-const translationRu = {
-  Sunday: 'Воскресенье',
-  Monday: 'Понедельник',
-  Tuesday: 'Вторник',
-  Wednesday: 'Среда',
-  Thursday: 'Четверг',
-  Friday: 'Пятница',
-  Saturday: 'Суббота',
-  January: 'января',
-  February: 'февраля',
-  March: 'марта',
-  April: 'апреля',
-  May: 'мая',
-  June: 'июня',
-  July: 'июля',
-  August: 'августа',
-  September: 'сентября',
-  October: 'октября',
-  November: 'ноября',
-  December: 'декабря',
-  font: '"Marmelad", sans-serif',
-  language: 'Язык',
-  english: 'Английский',
-  russian: 'Русский',
-  time: 'Время',
-};
-
-const translationEn = {
-  Sunday: 'Sunday',
-  Monday: 'Monday',
-  Tuesday: 'Tuesday',
-  Wednesday: 'Wednesday',
-  Thursday: 'Thursday',
-  Friday: 'Friday',
-  Saturday: 'Saturday',
-  January: 'January',
-  February: 'February',
-  March: 'March',
-  April: 'April',
-  May: 'May',
-  June: 'June',
-  July: 'July',
-  August: 'August',
-  September: 'September',
-  October: 'October',
-  November: 'November',
-  December: 'December',
-  font: '"Quicksand", sans-serif',
-  language: 'Language',
-  english: 'English',
-  russian: 'Russian',
-  time: 'Time',
-};
-
-/* export const addText = () => {
-  timeSettingSpan.innerHTML = `${i18next.t('time')}`;
-  greetingsSettingSpan.innerHTML = `${i18next.t('greeting')}`;
-  focusSettingSpan.innerHTML = `${i18next.t('focusSet')}`;
-  audiosettingSpan.innerHTML = `${i18next.t('audio')} `;
-  weatherSettingSpan.innerHTML = `${i18next.t('weather')}`;
-  quotesSettingSpan.innerHTML = `${i18next.t('quotesSet')}`;
-  enSetting.innerHTML = `${i18next.t('english')}`;
-  ruSetting.innerHTML = `${i18next.t('russian')}`;
-  langSetting.innerHTML = `${i18next.t('language')}`;
-  showSetting.innerHTML = `${i18next.t('show')}`;
-}; */
-
-function init() {
-  showTime();
-}
-
-i18next.init({
-  lng: 'en',
-  resources: {
-    en: {
-      translation: translationEn
-    },
-    ru: {
-      translation: translationRu
-    }
-  }
-});
-
-const listnerLang = () => {
-  sliderLang.addEventListener('click', () => {
-    if (sliderLang.checked === true) {
-      i18next.changeLanguage('ru', () => {
-        init();
-      });
-    } else {
-      i18next.changeLanguage('en', () => {
-        init();
-      });
-    }
-  });
-};
-listnerLang();
