+++ js-npm-task/package.json
@@ -4,9 +4,8 @@
   "description": "npm-task",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build_sass": "sass style.scss style.css",
-    "watch_sass": "sass style.scss style.css -w"
+    "build": "webpack --mode production",
+    "server": "webpack-dev-server --mode development"
   },
   "keywords": [],
   "author": "Tsakunova Tatsiana",
@@ -15,6 +14,12 @@
     "i18next": "^21.6.14"
   },
   "devDependencies": {
-    "sass": "^1.49.11"
+    "css-loader": "^6.7.1",
+    "html-webpack-plugin": "^5.5.0",
+    "sass-loader": "^12.6.0",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.71.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.7.4"
   }
 }
