+++ js-npm-task/src/index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>NPM task</title>
+  <link rel="preconnect" href="https://fonts.googleapis.com">
+  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+  <link href="https://fonts.googleapis.com/css2?family=Marmelad&family=Quicksand&display=swap" rel="stylesheet">
+</head>
+<body>
+  <section class="card">
+    <div class="switch-lang">
+        <p class="lang"></p>
+        <label id="switch" class="switch">
+          <input type="checkbox" id="slider">
+          <span class="slider round"></span>
+        </label>
+      </div>
+      <div id="time"></div>
+      <div class="today"></div>
+      <div id="week"></div>
+  </section>
+</body>
+</html>
