+++ js-npm-task/webpack.config.js
@@ -1,0 +1,41 @@
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const path = require('path')
+
+module.exports = {
+    entry: {
+        main: path.resolve(__dirname, './src/index.js'),
+    },
+    output: {
+      path: path.resolve(__dirname, './dist'),
+      filename: '[name].bundle.js',
+  },
+  devServer: {
+    historyApiFallback: true,
+    static: path.resolve(__dirname, './dist'),
+    open: true,
+    compress: true,
+    hot: true,
+    port: 8080,
+  },
+  plugins: [
+    new HtmlWebpackPlugin({
+      title: 'Momentum',
+      template: path.resolve(__dirname, './src/index.html'), // шаблон
+      filename: 'index.html', // название выходного файла
+    })],
+  module: {
+    rules: [
+      {
+        test: /\.s[ac]ss$/i,
+        use: [
+          // Creates `style` nodes from JS strings
+          "style-loader",
+          // Translates CSS into CommonJS
+          "css-loader",
+          // Compiles Sass to CSS
+          "sass-loader",
+        ],
+      },
+    ],
+  }
+}
